Changed storage system of enemy ships.
Enemy ships storage system's is now a list of lists of ships, each
sublist representing a column in the formation.
Changed EnemyShipFormation to adjust to the storage change.
Added logging to EnemyShipFormation.
Added logger to GameScreen to notify when the player ship receives a
hit.
Adjuste logger level to info.